esp32_webserver/simulation_system/requirements.txt
rel
websocket-client
matplotlib